@@ -1,11 +1,7 @@
 <template>
   <div class="col-md-12">
     <div class="card card-container">
-      <img
-        id="profile-img"
-        src="//ssl.gstatic.com/accounts/ui/avatar_2x.png"
-        class="profile-img-card"
-      />
+      <img id="profile-img" src="//ssl.gstatic.com/accounts/ui/avatar_2x.png" class="profile-img-card" />
       <Form @submit="handleLogin" :validation-schema="schema">
         <div class="form-group">
           <label for="email">Email</label>
@@ -21,10 +17,7 @@
         </div>
         <div class="form-group">
           <button class="btn btn-primary btn-block" :disabled="loading">
-            <span
-              v-show="loading"
-              class="spinner-border spinner-border-sm"
-            ></span>
+            <span v-show="loading" class="spinner-border spinner-border-sm"></span>
             <span>Login</span>
           </button>
         </div>
@@ -34,6 +27,11 @@
           </div>
         </div>
       </Form>
+      <div class="form-group">
+        <button class="btn btn-primary btn-block" @click="handleForgetPassword">
+          <span>Forget Password</span>
+        </button>
+      </div>
     </div>
   </div>
 </template>
@@ -48,7 +46,7 @@ export default {
     ErrorMessage,
   },
   data() {
-    var errors = {"email" : "", "password": ""};
+    var errors = { "email": "", "password": "" };
     const schema = yup.object().shape({
       email: yup
         .string()
@@ -79,6 +77,9 @@ export default {
     }
   },
   methods: {
+    handleForgetPassword() {
+      this.$router.push("/forget-password");
+    },
     handleLogin(user) {
       this.loading = true;
       this.$store.dispatch("auth/login", user).then(
@@ -94,10 +95,10 @@ export default {
             error.message ||
             error.toString();
 
-          if(error.response.data.error){
+          if (error.response.data.error) {
             this.errors['email'] = error.response.data.error
-          } else if(error.response.data.violations){
-            error.response.data.violations.forEach(element => 
+          } else if (error.response.data.violations) {
+            error.response.data.violations.forEach(element =>
               this.errors[element.propertyPath] = element.message
             );
           }
@@ -109,33 +110,36 @@ export default {
 </script>
 
 <style scoped>
-    label {
-    display: block;
-    margin-top: 10px;
-    }
-    .card-container.card {
-    max-width: 350px !important;
-    padding: 40px 40px;
-    }
-    .card {
-    background-color: #f7f7f7;
-    padding: 20px 25px 30px;
-    margin: 0 auto 25px;
-    margin-top: 50px;
-    -moz-border-radius: 2px;
-    -webkit-border-radius: 2px;
-    border-radius: 2px;
-    -moz-box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
-    -webkit-box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
-    box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
-    }
-    .profile-img-card {
-    width: 96px;
-    height: 96px;
-    margin: 0 auto 10px;
-    display: block;
-    -moz-border-radius: 50%;
-    -webkit-border-radius: 50%;
-    border-radius: 50%;
-    }
+label {
+  display: block;
+  margin-top: 10px;
+}
+
+.card-container.card {
+  max-width: 350px !important;
+  padding: 40px 40px;
+}
+
+.card {
+  background-color: #f7f7f7;
+  padding: 20px 25px 30px;
+  margin: 0 auto 25px;
+  margin-top: 50px;
+  -moz-border-radius: 2px;
+  -webkit-border-radius: 2px;
+  border-radius: 2px;
+  -moz-box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
+  -webkit-box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
+  box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
+}
+
+.profile-img-card {
+  width: 96px;
+  height: 96px;
+  margin: 0 auto 10px;
+  display: block;
+  -moz-border-radius: 50%;
+  -webkit-border-radius: 50%;
+  border-radius: 50%;
+}
 </style>
